requirements.txt
amqp==2.1.4
appdirs==1.4.2
billiard==3.5.0.2
boto3==1.4.4
botocore==1.5.27
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.10.5
djangorestframework==3.7.3
docutils==0.13.1
enum34==1.1.6
gunicorn==19.6.0
idna==2.6
ipdb==0.10.3
# ipython
# ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.2
kombu==4.0.2
lxml==4.2.3
mysqlclient==1.3.9
natsort==5.1.1
packaging==16.8
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
plivo==0.11.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyblake2==0.9.3
pycrypto==2.6.1
Pygments==2.2.0
PyMySQL==0.7.9
pyparsing==2.1.10
python-dateutil==2.6.0
python-memcached==1.58
pytz==2016.10
redis==2.10.5
requests==2.18.4
s3transfer==0.1.10
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.3
wcwidth==0.1.7
pyzmq==17.0.0
python-crontab==2.2.8
hashids==1.2.0
